refactor:extract achievement calculation logic and separate query/sync services
