+++ src/main/java/com/fesi/deadlinemate/domain/todo/service/TodoService.java
@@ -1,5 +1,6 @@
 package com.fesi.deadlinemate.domain.todo.service;
 
+import com.fesi.deadlinemate.domain.achievement.service.AchievementQueryService;
 import com.fesi.deadlinemate.domain.achievement.service.AchievementService;
 import com.fesi.deadlinemate.domain.gathering.entity.Gathering;
 import com.fesi.deadlinemate.domain.gathering.entity.GatheringMember;
